Extracted code to ArmyListRepositorySqlite.cs
